Master

+++ iina/MPVChapter.swift
@@ -3,13 +3,13 @@
 //  iina
 //
 //  Created by lhc on 29/8/16.
-//  Copyright © 2016年 lhc. All rights reserved.
+//  Copyright © 2016 lhc. All rights reserved.
 //
 
 import Cocoa
 
 class MPVChapter: NSObject {
-  
+
   private var privTitle: String?
   var title: String {
     get {
@@ -18,8 +18,8 @@ class MPVChapter: NSObject {
   }
   var time: VideoTime
   var index: Int
-  
-  init(title: String?, startTime: Int, index: Int) {
+
+  init(title: String?, startTime: Double, index: Int) {
     self.privTitle = title
     self.time = VideoTime(startTime)
     self.index = index
